pixart-rf: Renaming file and integrate function
