+++ plugins/pixart-rf/fu-plugin-pixart-rf.c
@@ -10,7 +10,7 @@
 
 
 #include "fu-pxi-device.h"
-#include "fu-pxi-wireless-device.h"
+#include "fu-pxi-dongle-receiver.h"
 #include "fu-pxi-firmware.h"
 
 void
@@ -20,6 +20,6 @@ fu_plugin_init (FuPlugin *plugin)
 	fu_plugin_set_build_hash (plugin, FU_BUILD_HASH);
 	fu_context_add_udev_subsystem (ctx, "hidraw");
 	fu_plugin_add_device_gtype (plugin, FU_TYPE_PXI_DEVICE);
-	fu_plugin_add_device_gtype (plugin, FU_TYPE_PXI_WIRELESS_DEVICE);
+	fu_plugin_add_device_gtype (plugin, FU_TYPE_PXI_DONGLE_RECEIVER);
 	fu_plugin_add_firmware_gtype (plugin, "pixart", FU_TYPE_PXI_FIRMWARE);
 }
