+++ plugins/pixart-rf/meson.build
@@ -11,9 +11,8 @@ shared_module('fu_plugin_pixart_rf',
     'fu-plugin-pixart-rf.c',
     'fu-pxi-device.c',
     'fu-pxi-common.c',
-    'fu-pxi-wireless-device.c',
-    'fu-pxi-wireless-module.c',
-    'fu-pxi-dongle-module.c',
+    'fu-pxi-dongle-receiver.c',
+    'fu-pxi-wireless-peripheral.c',
     'fu-pxi-firmware.c',      # fuzzing
   ],
   include_directories : [
